// src/app/generated_services/model/showTransactionCategoryDTO.ts
/**
 * BjjClutch Backend API
 * API for managing students, belts, graduation requirements, notices, frequencies and user authentication with Keycloak
 *
 * Contact: marcosturossi@gmail.com.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ShowTransactionCategoryDTO { 
    id?: string;
    name?: string | null;
    isActive?: boolean;
    createdAt?: string;
}